% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_4_shinyapp.R
\name{expand_bbox}
\alias{expand_bbox}
\title{Expand the bounding box of the polygon to show some context}
\usage{
expand_bbox(bbox, zoom_factor = 1.2)
}
\arguments{
\item{bbox}{coordinates given by \code{\link[sf]{st_bbox}}}

\item{zoom_factor}{numeric, multiplier ... default is 1.2}
}
\value{
zoomed coordinates
}
\description{
Expand the bounding box of the polygon to show some context
}
