% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_4_shinyapp.R
\name{myLeg}
\alias{myLeg}
\title{Create Custom Legend for Map}
\usage{
myLeg(vals, palette = "viridis")
}
\arguments{
\item{vals}{vector of values to create palette. Usually the
variable being mapped.}

\item{palette}{colors or color function that values will be
mapped to. This is passed directly to the
palette argument of \code{\link[leaflet:colorNumeric]{colorQuantile}}}
}
\value{
a ggplot object acting as a legend
}
\description{
Create Custom Legend for Map
}
