% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_4_shinyapp.R
\name{myPal}
\alias{myPal}
\title{Create the leaflet color palette function and choose the
optimal* number of breaks.}
\usage{
myPal(vals, palette = "viridis")
}
\arguments{
\item{vals}{vector of values to create palette. Usually the
variable being mapped.}

\item{palette}{colors or color function that values will be
mapped to. This is passed directly to the
palette argument of \code{\link[leaflet:colorNumeric]{colorQuantile}}}
}
\value{
a palette function
}
\description{
*In this case the optimal value is the value that allows
leaflet to work.
}
