% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_cdt.R
\name{runCDT}
\alias{runCDT}
\title{Run the current version of the contextual data tool}
\usage{
runCDT()
}
\value{
null
}
\description{
Include a legend
Consider how breaks are made-
  consistency across measures?
  natural breaks?
}
